+++ database/migrations/2019_02_22_151006_create_orcamento_produto_table.php
@@ -11,9 +11,9 @@ class CreateOrcamentoProdutoTable extends Migration {
 			$table->increments('id');
 			$table->integer('orcamento_id')->unsigned();
 			$table->integer('produto_id')->unsigned();
-			$table->integer('preco_unitario');
+			$table->float('preco_unitario');
 			$table->integer('quantidade');
-			$table->integer('preco_total');
+			$table->float('preco_total');
 		});
 	}
